Break apart profile into components
